refactor(shared.domain): split domain models into separate files; move BalanceItem, PriceHistory, SymbolPrice; replace BigDecimal with Double for commonMain; remove domain WalletState to use state.WalletState; update repos/usecases/viewmodels and tests

+++ androidApp/app/src/main/java/com/cryptx/cryptx/ui/components/AssetItem.kt
@@ -16,15 +16,14 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.cryptx.cryptx.ui.theme.*
-import java.math.BigDecimal
 import java.text.NumberFormat
 import java.util.Locale
 
 data class AssetItemData(
     val symbol: String,
     val name: String,
-    val quantity: BigDecimal,
-    val value: BigDecimal,
+    val quantity: Double,
+    val value: Double,
     val change: Double
 )
 
@@ -148,7 +147,7 @@ fun MiniChart(
     }
 }
 
-private fun formatCurrency(amount: BigDecimal): String {
+private fun formatCurrency(amount: Double): String {
     val formatter = NumberFormat.getCurrencyInstance(Locale.US)
     return formatter.format(amount)
 }
